Revert "Issue #40 : Form Validation rules added"
